Use bouncy castle provider for API 19+ support
