+++ app/build.gradle
@@ -42,7 +42,6 @@ dependencies {
     implementation "com.github.davidmigloz:number-keyboard:$numberKeyBoardVersion"
     implementation "com.brandongogetap:stickyheaders:$stickyheadersVersion"
     implementation 'com.android.support:multidex:1.0.3'
-    implementation 'com.madgag.spongycastle:core:1.58.0.0'
     testImplementation "junit:junit:$junitVersion"
     androidTestImplementation "com.android.support.test:runner:$testRunnerVersion"
     androidTestImplementation "com.android.support.test.espresso:espresso-core:$espressoVersion"
